Split private extensions from bridge core
