+++ package.json
@@ -31,7 +31,7 @@
   },
   "scripts": {
     "feishu-bot": "node ./bin/codex-im.js",
-    "check": "node --check ./src/index.js && node --check src/app/feishu-bot-runtime.js && node --check ./src/app/dispatcher.js && node --check ./src/app/card-action-config.js && node --check ./src/app/command-dispatcher.js && node --check ./src/app/codex-event-service.js && node --check src/infra/codex/message-utils.js && node --check ./src/infra/feishu/client-adapter.js && node --check ./src/infra/feishu/reaction-repo.js && node --check ./src/infra/codex/rpc-client.js && node --check ./src/infra/storage/session-store.js && node --check ./src/infra/config/config.js && node --check ./src/domain/approval/approval-service.js && node --check ./src/domain/approval/approval-policy.js && node --check ./src/domain/thread/thread-service.js && node --check ./src/domain/workspace/workspace-service.js && node --check ./src/domain/memory-bridge/memory-bridge-service.js && node --check ./src/domain/hub/hub-service.js && node --check ./src/domain/session/binding-context.js && node --check ./src/shared/workspace-paths.js && node --check ./src/shared/command-parsing.js && node --check ./src/shared/constants.js && node --check ./src/shared/errors.js && node --check ./src/presentation/card/builders.js && node --check ./src/presentation/card/card-service.js && node --check ./src/presentation/message/normalizers.js"
+    "check": "find src bin scripts -name '*.js' -print0 | xargs -0 -n1 node --check"
   },
   "dependencies": {
     "@larksuiteoapi/node-sdk": "^1.54.0",
