+++ src/app/feishu-bot-runtime.js
@@ -48,8 +48,8 @@ const approvalRuntime = require("../domain/approval/approval-service");
 const runtimeState = require("../domain/session/binding-context");
 const threadRuntime = require("../domain/thread/thread-service");
 const workspaceRuntime = require("../domain/workspace/workspace-service");
-const memoryBridgeRuntime = require("../domain/memory-bridge/memory-bridge-service");
-const hubRuntime = require("../domain/hub/hub-service");
+const memoryBridgeRuntime = require("../private/extensions/memory-bridge/memory-bridge-service");
+const hubRuntime = require("../private/extensions/hub/hub-service");
 const eventsRuntime = require("./codex-event-service");
 const approvalPolicyRuntime = require("../domain/approval/approval-policy");
 const appDispatcher = require("./dispatcher");
