+++ src/private/extensions/hub/hub-service.js
